organize discord client code
