+++ lib/collabland/connectSpace.ts
@@ -1,8 +1,8 @@
 import { InvalidInputError } from '@charmverse/core/errors';
 import { prisma } from '@charmverse/core/prisma-client';
 
-import { updateMemberRolesFromDiscord } from 'lib/discord/updateMemberRolesFromDiscord';
-import { upsertSpaceRolesFromDiscord } from 'lib/discord/upsertSpaceRolesFromDiscord';
+import { updateMemberRolesFromDiscord } from 'lib/discord/collabland/updateMemberRolesFromDiscord';
+import { upsertSpaceRolesFromDiscord } from 'lib/discord/collabland/upsertSpaceRolesFromDiscord';
 import { mapSpace } from 'lib/public-api/createWorkspaceApi';
 import { decryptData } from 'lib/utilities/dataEncryption';
