+++ lib/collabland/findOrCreateCollablandRoles.ts
@@ -1,7 +1,7 @@
 import type { Role } from '@charmverse/core/prisma';
 import { prisma } from '@charmverse/core/prisma-client';
 
-import { syncDiscordRoles } from 'lib/discord/syncDiscordRoles';
+import { syncDiscordRoles } from 'lib/discord/collabland/syncDiscordRoles';
 
 type RolesRecord = Record<string, Role | null>;
