+++ lib/discord/client/getDiscordAccount.ts
