+++ lib/discord/client/handleDiscordResponse.ts
