+++ lib/discord/collabland/checkDiscordGate.ts
@@ -1,7 +1,7 @@
 import { prisma } from '@charmverse/core/prisma-client';
 
+import { verifyDiscordGateForSpace } from 'lib/discord/collabland/verifyDiscordGateForSpace';
 import type { CheckDiscordGateResult } from 'lib/discord/interface';
-import { verifyDiscordGateForSpace } from 'lib/discord/verifyDiscordGateForSpace';
 import { InvalidInputError } from 'lib/utilities/errors';
 
 type Props = {
