+++ lib/discord/collabland/verifyDiscordGateForSpace.ts
