+++ lib/tokenGates/applyTokenGates.ts
@@ -2,7 +2,7 @@ import type { Role, Space } from '@charmverse/core/prisma';
 import { prisma } from '@charmverse/core/prisma-client';
 import { v4 } from 'uuid';
 
-import { applyDiscordGate } from 'lib/discord/applyDiscordGate';
+import { applyDiscordGate } from 'lib/discord/collabland/applyDiscordGate';
 import { checkUserSpaceBanStatus } from 'lib/members/checkUserSpaceBanStatus';
 import { trackUserAction } from 'lib/metrics/mixpanel/trackUserAction';
 import { updateTrackUserProfileById } from 'lib/metrics/mixpanel/updateTrackUserProfileById';
